// client/src/app/models/product.ts
export type Product = {
  id: number;
  name: string;
  description: string;
  price: number;
  pictureUrl: string;
  type: string;
  brand: string;
  quantityInStock: number;
}

// export interface Product {
//   id: number;
//   name: string;
//   description: string;
//   price: number;
//   pictureUrl: string;
//   type: string;
//   brand: string;
//   quantityInStock: number;
// }